examples/mccl_ddp_train/pyproject.toml
[project]
name = "mccl-ddp-example"
version = "0.1.0"
description = "Minimal distributed training example using PyTorch MPS + MCCL"
requires-python = ">=3.11"
dependencies = [
    "torch>=2.5.0",
    "mccl>=0.3.0",
    "numpy>=1.20",
    "pyyaml>=6.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"
